refactor(semantic): split ast_utility into display/types/decls
