+++ tests/snap/document_symbol/members.cpp
@@ -1,0 +1,29 @@
+// Detail rendering for class members: constructors (return type stripped),
+// destructors, operators, conversion functions, statics, bitfields, and
+// out-of-line definitions.
+
+namespace members {
+
+struct Widget {
+    Widget();
+    explicit Widget(int size);
+    ~Widget();
+
+    Widget& operator=(const Widget& other);
+    bool operator==(const Widget& other) const;
+    operator bool() const;
+
+    static int instances();
+
+    int size;
+    unsigned bits : 3;
+    const char* name = "widget";
+};
+
+Widget::Widget(int size) : size(size), bits(0) {}
+
+int Widget::instances() {
+    return 0;
+}
+
+}  // namespace members
