+++ tests/snap/document_symbol/templates.snap.yml
@@ -1,0 +1,17 @@
+---
+created_at: 2026-07-30
+input_file: templates.cpp
+---
+- { name: "templates", kind: Namespace, range: "5:0-52:1", selection_range: "5:10-5:19" }
+-   { name: "Box", kind: Struct, range: "8:0-12:1", selection_range: "8:7-8:10", detail: "template struct" }
+-     { name: "value", kind: Field, range: "9:4-9:11", selection_range: "9:6-9:11", detail: "T" }
+-     { name: "reset", kind: Method, range: "11:4-11:16", selection_range: "11:9-11:14", detail: "void ()" }
+-   { name: "Box<T>::reset", kind: Method, range: "14:0-15:23", selection_range: "15:13-15:18", detail: "void ()" }
+-   { name: "pointee", kind: Field, range: "22:4-22:14", selection_range: "22:7-22:14", detail: "T *" }
+-   { name: "zero", kind: Function, range: "26:0-28:1", selection_range: "26:2-26:6", detail: "template T ()" }
+-   { name: "zero<int>", kind: Function, range: "30:0-31:15", selection_range: "31:4-31:8", detail: "int ()" }
+-   { name: "pi", kind: Variable, range: "34:0-34:27", selection_range: "34:12-34:14", detail: "template const T" }
+-   { name: "Small", kind: TypeParameter, range: "36:0-37:30", selection_range: "37:8-37:13", detail: "concept" }
+-   { name: "Guide", kind: Struct, range: "39:0-42:1", selection_range: "39:7-39:12", detail: "struct" }
+-     { name: "Guide", kind: Method, range: "41:4-41:16", selection_range: "41:4-41:9", detail: "template (T)" }
+-   { name: "T", kind: Variable, range: "45:0-45:9", selection_range: "45:6-45:7", detail: "template Guide" }
