+++ tests/snap/inlay_hint/param_hints.snap.yml
@@ -1,0 +1,24 @@
+---
+created_at: 2026-07-30
+input_file: param_hints.cpp
+---
+- { pos: "35:9", kind: Parameter, label: "width:", padding_right: true }
+- { pos: "35:13", kind: Parameter, label: "height:", padding_right: true }
+- { pos: "47:16", kind: Parameter, label: "height:", padding_right: true }
+- { pos: "71:23", kind: Parameter, label: "timeout_millis:", padding_right: true }
+- { pos: "82:12", kind: Parameter, label: "sink:", padding_right: true }
+- { pos: "94:11", kind: Parameter, label: "&value:", padding_right: true }
+- { pos: "95:12", kind: Parameter, label: "value:", padding_right: true }
+- { pos: "110:9", kind: Parameter, label: "first:", padding_right: true }
+- { pos: "110:12", kind: Parameter, label: "second:", padding_right: true }
+- { pos: "120:11", kind: Parameter, label: "width:", padding_right: true }
+- { pos: "120:16", kind: Parameter, label: "height:", padding_right: true }
+- { pos: "132:9", kind: Parameter, label: "value:", padding_right: true }
+- { pos: "132:12", kind: Parameter, label: "count:", padding_right: true }
+- { pos: "150:14", kind: Parameter, label: "a:", padding_right: true }
+- { pos: "150:17", kind: Parameter, label: "b:", padding_right: true }
+- { pos: "150:20", kind: Parameter, label: "c:", padding_right: true }
+- { pos: "163:12", kind: Parameter, label: "x:", padding_right: true }
+- { pos: "163:15", kind: Parameter, label: "y:", padding_right: true }
+- { pos: "165:12", kind: Parameter, label: "x:", padding_right: true }
+- { pos: "165:15", kind: Parameter, label: "y:", padding_right: true }
