+++ tests/snap/inlay_hint/type_hints.snap.yml
@@ -1,0 +1,17 @@
+---
+created_at: 2026-07-30
+input_file: type_hints.cpp
+---
+- { pos: "10:14", kind: Type, label: ": int" }
+- { pos: "11:19", kind: Type, label: ": const int &" }
+- { pos: "12:13", kind: Type, label: ": int *" }
+- { pos: "24:14", kind: Type, label: ": Integer" }
+- { pos: "39:14", kind: Type, label: ": Compact" }
+- { pos: "54:11", kind: Type, label: ": int" }
+- { pos: "54:14", kind: Type, label: ": float" }
+- { pos: "62:17", kind: Type, label: ": (lambda)" }
+- { pos: "62:29", kind: Type, label: "-> int" }
+- { pos: "71:16", kind: Type, label: ": int" }
+- { pos: "75:19", kind: Parameter, label: "x:", padding_right: true }
+- { pos: "81:13", kind: Type, label: "-> int" }
+- { pos: "85:18", kind: Type, label: "-> int &" }
