+++ tests/unit/feature/inlay_hint_tests.cpp
@@ -19,12 +19,14 @@ TEST_SUITE(inlay_hint, Tester) {
 std::vector<protocol::InlayHint> hints;
 llvm::DenseMap<std::uint32_t, protocol::InlayHint> hints_map;
 
-void run(llvm::StringRef code, std::source_location location = std::source_location::current()) {
+void run(llvm::StringRef code,
+         const feature::InlayHintsOptions& options = {},
+         std::source_location location = std::source_location::current()) {
     add_main("main.cpp", code);
     ASSERT_TRUE(compile_with_pch("-std=c++23"));
 
     LocalSourceRange range = LocalSourceRange(0, unit->interested_content().size());
-    hints = feature::inlay_hints(*unit, range, {}, feature::PositionEncoding::UTF8);
+    hints = feature::inlay_hints(*unit, range, options, feature::PositionEncoding::UTF8);
 
     hints_map.clear();
     auto content = unit->interested_content();
@@ -901,12 +903,12 @@ TEST_CASE(Designators, skip = true) {
     EXPECT_HINT("0", ".b=");
 };
 
-TEST_CASE(BlockEnd, skip = true) {
+TEST_CASE(BlockEnd) {
     // Functions
     run(R"c(
             int foo() {
                 return 41;
-            §(0)}
+            }§(0)
 
             template<int X>
             int bar() {
@@ -915,7 +917,7 @@ TEST_CASE(BlockEnd, skip = true) {
                     return X;
                 };
                 return f();
-            §(1)}
+            }§(1)
 
             // No hint because this isn't a definition
             int buz();
@@ -923,12 +925,13 @@ TEST_CASE(BlockEnd, skip = true) {
             struct S{};
             bool operator==(S, S) {
                 return true;
-            §(2)}
-        )c");
+            }§(2)
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(3);
-    EXPECT_HINT("0", " // foo");
-    EXPECT_HINT("1", " // bar");
-    EXPECT_HINT("2", " // operator==");
+    EXPECT_HINT("0", "// foo");
+    EXPECT_HINT("1", "// bar");
+    EXPECT_HINT("2", "// operator==");
 
     // Methods
     run(R"c(
@@ -937,17 +940,17 @@ TEST_CASE(BlockEnd, skip = true) {
                 Test() = default;
 
                 ~Test() {
-                §(0)}
+                }§(0)
 
                 void method1() {
-                §(1)}
+                }§(1)
 
                 // No hint because this isn't a definition
                 void method2();
 
                 template <typename T>
                 void method3() {
-                §(2)}
+                }§(2)
 
                 // No hint because this isn't a definition
                 template <typename T>
@@ -955,69 +958,72 @@ TEST_CASE(BlockEnd, skip = true) {
 
                 Test operator+(int) const {
                     return *this;
-                §(3)}
+                }§(3)
 
                 operator bool() const {
                     return true;
-                §(4)}
+                }§(4)
 
                 // No hint because there's no function body
                 operator int() const = delete;
             } x;
 
             void Test::method2() {
-            §(5)}
+            }§(5)
 
             template <typename T>
             void Test::method4() {
-            §(6)}
-        )c");
+            }§(6)
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(7);
-    EXPECT_HINT("0", " // ~Test");
-    EXPECT_HINT("1", " // method1");
-    EXPECT_HINT("2", " // method3");
-    EXPECT_HINT("3", " // operator+");
-    EXPECT_HINT("4", " // operator bool");
-    EXPECT_HINT("5", " // Test::method2");
-    EXPECT_HINT("6", " // Test::method4");
+    EXPECT_HINT("0", "// ~Test");
+    EXPECT_HINT("1", "// method1");
+    EXPECT_HINT("2", "// method3");
+    EXPECT_HINT("3", "// operator+");
+    EXPECT_HINT("4", "// operator bool");
+    EXPECT_HINT("5", "// Test::method2");
+    EXPECT_HINT("6", "// Test::method4");
 
     // Namespaces
     run(R"c(
             namespace {
                 void foo();
-            §(0)}
+            }§(0)
 
             namespace ns {
                 void bar();
-            §(1)}
-        )c");
+            }§(1)
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(2);
-    EXPECT_HINT("0", " // namespace");
-    EXPECT_HINT("1", " // namespace ns");
+    EXPECT_HINT("0", "// namespace");
+    EXPECT_HINT("1", "// namespace ns");
 
     // Types
     run(R"c(
             struct S {
-            §(0)};
+            };§(0)
 
             class C {
-            §(1)};
+            };§(1)
 
             union U {
-            §(2)};
+            };§(2)
 
             enum E1 {
-            §(3)};
+            };§(3)
 
             enum class E2 {
-            §(4)};
-        )c");
+            };§(4)
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(5);
-    EXPECT_HINT("0", " // struct S");
-    EXPECT_HINT("1", " // class C");
-    EXPECT_HINT("2", " // union U");
-    EXPECT_HINT("3", " // enum E1");
-    EXPECT_HINT("4", " // enum class E2");
+    EXPECT_HINT("0", "// struct S");
+    EXPECT_HINT("1", "// class C");
+    EXPECT_HINT("2", "// union U");
+    EXPECT_HINT("3", "// enum E1");
+    EXPECT_HINT("4", "// enum class E2");
 
     // If statements
     run(R"c(
@@ -1026,37 +1032,51 @@ TEST_CASE(BlockEnd, skip = true) {
                     ;
 
                 if (cond) {
-                §(0)}
+                }§(0)
 
                 if (cond) {
                 } else {
-                §(1)}
+                }§(1)
 
                 if (cond) {
                 } else if (!cond) {
-                §(2)}
+                }§(2)
 
                 if (cond) {
                 } else {
                     if (!cond) {
-                    §(3)}
-                §(4)}
+                    }§(3)
+                }§(4)
 
                 if (auto X = cond) {
-                §(5)}
+                }§(5)
 
                 if (int i = 0; i > 10) {
-                §(6)}
+                }§(6)
             }
-        )c");
-    EXPECT_SIZE(7);
-    EXPECT_HINT("0", " // if cond");
-    EXPECT_HINT("1", " // if cond");
-    EXPECT_HINT("2", " // if");
-    EXPECT_HINT("3", " // if !cond");
-    EXPECT_HINT("4", " // if cond");
-    EXPECT_HINT("5", " // if X");
-    EXPECT_HINT("6", " // if i > 10");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(8);
+    EXPECT_HINT("0", "// if cond");
+    EXPECT_HINT("1", "// if cond");
+    EXPECT_HINT("2", "// if");
+    EXPECT_HINT("3", "// if !cond");
+    EXPECT_HINT("4", "// if cond");
+    EXPECT_HINT("5", "// if X");
+    EXPECT_HINT("6", "// if i > 10");
+
+    // `if consteval` has no condition and hints as plain "// if".
+    run(R"c(
+            void foo() {
+                if consteval {
+                    int x = 1;
+                }§(0)
+            }§(1)
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(2);
+    EXPECT_HINT("0", "// if");
+    EXPECT_HINT("1", "// foo");
 
     // Loops
     run(R"c(
@@ -1065,64 +1085,68 @@ TEST_CASE(BlockEnd, skip = true) {
                     ;
 
                 while (true) {
-                §(0)}
+                }§(0)
 
                 do {
                 } while (true);
 
                 for (;true;) {
-                §(1)}
+                }§(1)
 
                 for (int I = 0; I < 10; ++I) {
-                §(2)}
+                }§(2)
 
                 int Vs[] = {1,2,3};
                 for (auto V : Vs) {
-                §(3)}
+                }§(3)
             }
-        )c");
-    EXPECT_SIZE(4);
-    EXPECT_HINT("0", " // while true");
-    EXPECT_HINT("1", " // for true");
-    EXPECT_HINT("2", " // for I");
-    EXPECT_HINT("3", " // for V");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(5);
+    EXPECT_HINT("0", "// while true");
+    EXPECT_HINT("1", "// for true");
+    EXPECT_HINT("2", "// for I");
+    EXPECT_HINT("3", "// for V");
 
     // Switch
     run(R"c(
             void foo(int I) {
                 switch (I) {
                     case 0: break;
-                §(0)}
+                }§(0)
             }
-        )c");
-    EXPECT_SIZE(1);
-    EXPECT_HINT("0", " // switch I");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(2);
+    EXPECT_HINT("0", "// switch I");
 
     // Print literals
     run(R"c(
+            #pragma clang diagnostic ignored "-Wliteral-conversion"
             void foo() {
                 while ("foo") {
-                §(0)}
+                }§(0)
 
                 while ("foo but this time it is very long") {
-                §(1)}
+                }§(1)
 
                 while (true) {
-                §(2)}
+                }§(2)
 
                 while (1) {
-                §(3)}
+                }§(3)
 
                 while (1.5) {
-                §(4)}
+                }§(4)
             }
-        )c");
-    EXPECT_SIZE(5);
-    EXPECT_HINT("0", " // while \"foo\"");
-    EXPECT_HINT("1", " // while \"foo but...\"");
-    EXPECT_HINT("2", " // while true");
-    EXPECT_HINT("3", " // while 1");
-    EXPECT_HINT("4", " // while 1.5");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(6);
+    EXPECT_HINT("0", "// while \"foo\"");
+    EXPECT_HINT("1", "// while \"foo but...\"");
+    EXPECT_HINT("2", "// while true");
+    EXPECT_HINT("3", "// while 1");
+    EXPECT_HINT("4", "// while 1.5");
 
     // Print refs
     run(R"c(
@@ -1136,23 +1160,24 @@ TEST_CASE(BlockEnd, skip = true) {
             }
             void foo() {
                 while (ns::Var) {
-                §(0)}
+                }§(0)
 
                 while (ns::func()) {
-                §(1)}
+                }§(1)
 
                 while (ns::S{}.Field) {
-                §(2)}
+                }§(2)
 
                 while (ns::S{}.method()) {
-                §(3)}
+                }§(3)
             }
-        )c");
-    EXPECT_SIZE(4);
-    EXPECT_HINT("0", " // while Var");
-    EXPECT_HINT("1", " // while func");
-    EXPECT_HINT("2", " // while Field");
-    EXPECT_HINT("3", " // while method");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(7);
+    EXPECT_HINT("0", "// while Var");
+    EXPECT_HINT("1", "// while func");
+    EXPECT_HINT("2", "// while Field");
+    EXPECT_HINT("3", "// while method");
 
     // Print conversions
     run(R"c(
@@ -1163,65 +1188,67 @@ TEST_CASE(BlockEnd, skip = true) {
             };
             void foo(int I) {
                 while (float(I)) {
-                §(0)}
+                }§(0)
 
                 while (S(I)) {
-                §(1)}
+                }§(1)
 
                 while (S(I, I)) {
-                §(2)}
+                }§(2)
             }
-        )c");
-    EXPECT_SIZE(3);
-    EXPECT_HINT("0", " // while float");
-    EXPECT_HINT("1", " // while S");
-    EXPECT_HINT("2", " // while S");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(5);
+    EXPECT_HINT("0", "// while float");
+    EXPECT_HINT("1", "// while S");
+    EXPECT_HINT("2", "// while S");
 
     // Print operators
     run(R"c(
             using Integer = int;
             void foo(Integer I) {
                 while(++I){
-                §(0)}
+                }§(0)
 
                 while(I++){
-                §(1)}
+                }§(1)
 
                 while(+(I + I)){
-                §(2)}
+                }§(2)
 
                 while(I < 0){
-                §(3)}
+                }§(3)
 
                 while((I + I) < I){
-                §(4)}
+                }§(4)
 
                 while(I < (I + I)){
-                §(5)}
+                }§(5)
 
                 while((I + I) < (I + I)){
-                §(6)}
+                }§(6)
             }
-        )c");
-    EXPECT_SIZE(7);
-    EXPECT_HINT("0", " // while ++I");
-    EXPECT_HINT("1", " // while I++");
-    EXPECT_HINT("2", " // while");
-    EXPECT_HINT("3", " // while I < 0");
-    EXPECT_HINT("4", " // while ... < I");
-    EXPECT_HINT("5", " // while I < ...");
-    EXPECT_HINT("6", " // while");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(8);
+    EXPECT_HINT("0", "// while ++I");
+    EXPECT_HINT("1", "// while I++");
+    EXPECT_HINT("2", "// while");
+    EXPECT_HINT("3", "// while I < 0");
+    EXPECT_HINT("4", "// while ... < I");
+    EXPECT_HINT("5", "// while I < ...");
+    EXPECT_HINT("6", "// while");
 
     // Trailing semicolon
     run(R"c(
             // The hint is placed after the trailing ';'
             struct S1 {
-            §(0)}  ;
+            }  ;§(0)
 
             // The hint is always placed in the same line with the closing '}'.
             // So in this case where ';' is missing, it is attached to '}'.
             struct S2 {
-            §(1)}
+            }§(1)
 
             ;
 
@@ -1236,19 +1263,20 @@ TEST_CASE(BlockEnd, skip = true) {
             // Rare case, but yes we'll have a hint here.
             struct {
                 int x;
-            §(2)}
+            }§(2)
 
             s2;
-        )c");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(3);
-    EXPECT_HINT("0", " // struct S1");
-    EXPECT_HINT("1", " // struct S2");
-    EXPECT_HINT("2", " // struct");
+    EXPECT_HINT("0", "// struct S1");
+    EXPECT_HINT("1", "// struct S2");
+    EXPECT_HINT("2", "// struct");
 
     // Trailing text
     run(R"c(
             struct S1 {
-            §(0)}      ;
+            }      ;§(0)
 
             // No hint for S2 because of the trailing comment
             struct S2 {
@@ -1257,15 +1285,16 @@ TEST_CASE(BlockEnd, skip = true) {
             struct S3 {
                 // No hint for S4 because of the trailing source code
                 struct S4 {
-                };§(1)};
+                };};§(1)
 
             // No hint for ns because of the trailing comment
             namespace ns {
             } // namespace ns
-        )c");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(2);
-    EXPECT_HINT("0", " // struct S1");
-    EXPECT_HINT("1", " // struct S3");
+    EXPECT_HINT("0", "// struct S1");
+    EXPECT_HINT("1", "// struct S3");
 
     // Macro
     run(R"c(
@@ -1273,14 +1302,15 @@ TEST_CASE(BlockEnd, skip = true) {
             #define RBRACE }
 
             DECL_STRUCT(S1)
-            §(0)};
+            };§(0)
 
             // No hint because we require a '}'
             DECL_STRUCT(S2)
             RBRACE;
-        )c");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
     EXPECT_SIZE(1);
-    EXPECT_HINT("0", " // struct S1");
+    EXPECT_HINT("0", "// struct S1");
 
     // Pointer to member function
     run(R"c(
@@ -1288,14 +1318,15 @@ TEST_CASE(BlockEnd, skip = true) {
             using Predicate = bool(A::*)();
             void foo(A* a, Predicate p) {
                 if ((a->*p)()) {
-                §(0)}
+                }§(0)
             }
-        )c");
-    EXPECT_SIZE(1);
-    EXPECT_HINT("0", " // if");
+        )c",
+        {.parameters = false, .deduced_types = false, .block_end = true});
+    EXPECT_SIZE(2);
+    EXPECT_HINT("0", "// if");
 };
 
-TEST_CASE(DefaultArguments, skip = true) {
+TEST_CASE(DefaultArguments) {
     // Smoke test
     run(R"c(
             int foo(int A = 4) { return A; }
@@ -1303,14 +1334,15 @@ TEST_CASE(DefaultArguments, skip = true) {
             int A = bar(§(1)2§(2));
 
             void baz(int = 5) { if (false) baz(§(3)); };
-        )c");
+        )c",
+        {.default_arguments = true});
     EXPECT_SIZE(4);
     EXPECT_HINT("0", "A: 4");
-    EXPECT_HINT("1", "A: ");
+    EXPECT_HINT("1", "A:");
     EXPECT_HINT("2", ", B: 1, C: foo()");
     EXPECT_HINT("3", "5");
 
-    // Without parameter names
+    // Packs and constructor calls
     run(R"c(
             struct Baz {
                 Baz(float a = 3 //
@@ -1324,19 +1356,20 @@ TEST_CASE(DefaultArguments, skip = true) {
                 ) {}
             };
 
-            int main() {
+            void use() {
                 Foo foo1(1§(1));
                 Foo foo2{2§(2)};
                 Foo foo3 = {3§(3)};
                 auto foo4 = Foo{4§(4)};
             }
-        )c");
-    EXPECT_SIZE(5);
-    EXPECT_HINT("0", "...");
-    EXPECT_HINT("1", ", Baz{}");
-    EXPECT_HINT("2", ", Baz{}");
-    EXPECT_HINT("3", ", Baz{}");
-    EXPECT_HINT("4", ", Baz{}");
+        )c",
+        {.default_arguments = true});
+    EXPECT_SIZE(6);
+    EXPECT_HINT("0", "a: ...");
+    EXPECT_HINT("1", ", baz: Baz{}");
+    EXPECT_HINT("2", ", baz: Baz{}");
+    EXPECT_HINT("3", ", baz: Baz{}");
+    EXPECT_HINT("4", ", baz: Baz{}");
 };
 
 // FIXME: flaky on some platforms, skip until root cause is identified.
